% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_theme.R
\docType{data}
\name{hc_theme_swi}
\alias{hc_theme_swi}
\title{Highchart(er) SWI theme}
\format{An object of class \code{hc_theme} of length 8.}
\usage{
hc_theme_swi
}
\description{
Highchart(er) SWI theme
}
\examples{
\dontrun{
require(highcharter)
hc_demo() \%>\% hc_add_theme(hc_theme_swi)
}
}
\seealso{
http://jkunst.com/highcharter/themes.html#create-themes
}
\keyword{datasets}

