% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlLink.R
\name{htmlLink}
\alias{htmlLink}
\title{Helper function to create html link}
\usage{
htmlLink(url, text, target = "_blank")
}
\arguments{
\item{url, }{a character string the target URL}

\item{text, }{a character string the text displayed as HTML link}

\item{target, }{a character valid attribute for target}
}
\description{
Helper function to create html link
}
\examples{
\dontrun{
htmlLink("http://www.swissinfo.ch/eng", "swi")
}
}
\seealso{
http://www.w3schools.com/tags/att_a_target.asp
}

