% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iframeresizer.R
\docType{data}
\name{iframeresizer}
\alias{iframeresizer}
\title{html code to call SWI iframeresizer}
\format{An object of class \code{character} of length 1.}
\usage{
iframeresizer
}
\description{
html code to call SWI iframeresizer
}
\examples{
\dontrun{
iframeresizer
}
}
\seealso{
https://github.com/davidjbradshaw/iframe-resizer
}
\keyword{datasets}

