% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_theme.R
\docType{data}
\name{hc_theme_swi}
\alias{hc_theme_swi}
\alias{style_swi_highcharter}
\title{Highchart(er) SWI theme}
\format{An object of class \code{hc_theme} of length 8.}
\usage{
hc_theme_swi

style_swi_highcharter()
}
\description{
Some helpers to get a SWI look for highcharter
}
\examples{
\dontrun{
require(highcharter)
data(diamonds, mpg, package = "ggplot2")
hchart(mpg, "scatter", x = displ, y = hwy, group = class) \%>\% hc_add_theme(hc_theme_swi)
}
style_swi_highcharter()
}
\seealso{
http://jkunst.com/highcharter/themes.html#create-themes
}
\keyword{datasets}

