% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiHTMLwidget.R
\name{swi_widget}
\alias{grab_widgetHTML}
\alias{swi_libWidget_overwrite}
\alias{swi_widget}
\title{Helper functions for htmlwidget}
\usage{
swi_widget(widget.html, output.html = "parset_swi.html", output = ".",
  source = "source:", author = " swissinfo.ch", h2 = "title",
  descr = "descriptive text", h3 = "subtitle", footer = "")

grab_widgetHTML(widget.html)

swi_libWidget_overwrite(dirPath = ".")
}
\arguments{
\item{widget.html, output.html}{character file path to the input widget html and the output reponsive html}

\item{output}{a path to a folder where the output html file and the js library folder will be saved}

\item{source, author, h2, descr, h3, footer}{characters}

\item{dir}{a path to a folder where all the js, css, ... assets on which the htmlwidget depends on}
}
\description{
Helper functions to create a container for htmlwidget with javascript library files
}
\examples{
\dontrun{
library(swiRcharts)
require(htmlwidgets)
require(parsetR)

widget <- parset(HairEyeColor, tension = 0.5)

tmp.dir <- tempdir()
widget.tmp.file <- paste0(tmp.dir, "/tmp.html")
widget.swi.file <- "output.html"
saveWidget(widget, file = widget.tmp.file, selfcontained = F, libdir = "js")
swi_widget(widget.tmp.file, widget.swi.file, output=tmp.dir)

viewer <- getOption("viewer")
viewer(widget.tmp.file)
viewer(paste0(tmp.dir, "/", widget.swi.file))
}
}

