% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiHTMLwidget.R
\name{swi_widget}
\alias{swi_widget}
\title{Helper functions for htmlwidget}
\usage{
swi_widget(widget.html, output.html = "parset_swi.html", output = ".",
  source = "source:", author = " swissinfo.ch", h2 = "title",
  descr = "descriptive text", h3 = "subtitle", footer = "")
}
\arguments{
\item{widget.html, output.html}{character file path to the input widget html and the output reponsive html}

\item{output}{a path to a folder where the output html file and the js library folder will be saved}

\item{source, author, h2, descr, h3, footer}{characters}
}
\description{
Helper functions to create a container for htmlwidget with javascript library files
}
\examples{
\dontrun{
TO DO !!!!!!
}
}

